Finished Tab and Tabs
Finished tabsjs, mapping over each item in the tabs array to make a Tab component for each one, sending down the strings for the corresponding string on the tabs array, the string for the selected tab in state on contentjs, and the function for onClick functionality as props. Ran into a blocker where I forgot the return keyword within the map(). Finished tab, creating the onClick event that takes the tab's string as an argument. Added the variable className for selectedTab. Finished all PropTypes for these components.