Full styled-components
Added style components to everything except the body. I found ways to do this on google search but did not have enough time to test them, couldn't get them working immediately. Nevertheless deleted the css file and everything looks fine except that the background color does not stretch to the full view of the page.